refactor: Update double to decimal in monetary values

+++ src/StorEsc.ApplicationServices/Interfaces/IRechargeApplicationService.cs
@@ -6,6 +6,6 @@ public interface IRechargeApplicationService
 {
     Task<bool> RechargeCustomerWalletAsync(
         string customerId,
-        double amount,
+        decimal amount,
         CreditCardDTO creditCardDTO);
 }
